:truck: Move plugin GUI into separate file
