+++ plugins/src/plugin.rs
@@ -1,17 +1,15 @@
-use crate::perm_tree::filesystem::FileWriteOptions;
-use crate::perm_tree::{
-    BusPermissions, ControlPermissions, CpuPermissions, ExternalPermissions, FileSystemPermissions,
-    InternalPermissions, PermTreeNode, PpuPermissions, RSnesPermissions,
-};
-use crate::permission::Permission;
-use crate::permission::helpers::AllOr;
+// uncomment below if code coverage is getting too low, as
+// it would be "fine" to count GUI code towards coverage
+// #[cfg(not(tarpaulin_include))]
+pub mod gui;
+
+use crate::perm_tree::{PermTreeNode, RSnesPermissions};
+use crate::plugin::gui::PluginPermRequest;
 
 use std::io::Read;
 use std::path::{Path, PathBuf};
 
 use common::snes_address::SnesAddress;
-use egui::text::LayoutJob;
-use egui::{CollapsingHeader, RichText, Style, TextFormat, WidgetText};
 use piccolo as picc;
 use piccolo::io as p_io;
 use std::fs;
@@ -322,179 +320,6 @@ impl Plugin {
         });
 
         lua.execute(&ex)
-    }
-}
-
-pub struct PluginPermRequest<'a> {
-    pub plugin: &'a Plugin,
-    pub allow_all: bool,
-
-    pub show_none: bool,
-}
-
-impl<'a> PluginPermRequest<'a> {
-    fn perm_label(perm: &impl Permission, name: &str) -> impl Into<WidgetText> {
-        let mut job = LayoutJob::default();
-        job.append(name, 0.0, Default::default());
-        if perm.is_none() {
-            job.append(
-                "none",
-                12.0,
-                TextFormat {
-                    italics: true,
-                    ..Default::default()
-                },
-            );
-        }
-        if perm.is_all() {
-            job.append(
-                "all",
-                12.0,
-                TextFormat {
-                    color: Style::default().visuals.strong_text_color(),
-                    italics: true,
-                    ..Default::default()
-                },
-            );
-        }
-
-        job
-    }
-    fn perm_collapsing_header(perm: &impl Permission, name: &str) -> CollapsingHeader {
-        CollapsingHeader::new(Self::perm_label(perm, name))
-    }
-    fn force_show_perm_collapsible<T: Permission>(
-        &self,
-        ui: &mut egui::Ui,
-        perm: &T,
-        label: &str,
-        draw_content: impl FnOnce(&mut egui::Ui, &T),
-    ) {
-        Self::perm_collapsing_header(perm, label)
-            .default_open(!perm.is_all() && !perm.is_none())
-            .show(ui, |ui| draw_content(ui, perm));
-    }
-    fn show_perm_collapsible<T: Permission>(
-        &self,
-        ui: &mut egui::Ui,
-        perm: &T,
-        label: &str,
-        draw_content: impl FnOnce(&mut egui::Ui, &T),
-    ) {
-        self.show_perm(ui, perm, |ui, perm| {
-            Self::force_show_perm_collapsible(&self, ui, perm, label, draw_content);
-        });
-    }
-    fn show_perm<T: Permission>(
-        &self,
-        ui: &mut egui::Ui,
-        perm: &T,
-        add_content: impl FnOnce(&mut egui::Ui, &T),
-    ) {
-        if !perm.is_none() || self.show_none {
-            add_content(ui, perm);
-        }
-    }
-    fn show_perm_bool(&self, ui: &mut egui::Ui, perm: bool, label: &str) {
-        self.show_perm(ui, &perm, |ui, perm| {
-            ui.label(Self::perm_label(perm, label));
-        });
-    }
-    pub fn show_gui(&mut self, ui: &mut egui::Ui) {
-        let close = |ui: &mut egui::Ui| {
-            ui.ctx().send_viewport_cmd(egui::ViewportCommand::Close);
-        };
-
-        ui.separator();
-
-        ui.label(RichText::new("Requested permissions").heading());
-        let RSnesPermissions { internal, external } = &self.plugin.table.perms;
-
-        self.force_show_perm_collapsible(ui, internal, "Internal", |ui, internal| {
-            // we intentionally destructure the stuct listing out all fields (without `..`),
-            // so that we get a compile error in case we forget to list a field in the
-            // destructure (and a warning if we write it just below but don't use it).
-            // This guarantees that we render all requested permissions in the GUI,
-            // which guarantees some security to the user (they at least know what is
-            // requested)
-            let InternalPermissions {
-                control,
-                cpu,
-                ppu,
-                bus,
-                input,
-            } = internal;
-
-            self.show_perm_collapsible(ui, cpu, "CPU", |ui, cpu| {
-                let CpuPermissions { registers } = cpu;
-                self.show_perm_bool(ui, *registers, "Registers");
-            });
-            self.show_perm_collapsible(ui, bus, "Bus", |ui, bus| {
-                let BusPermissions { read, write } = bus;
-                self.show_perm_bool(ui, *read, "Read");
-                self.show_perm_bool(ui, *write, "Write");
-            });
-            self.show_perm_collapsible(ui, ppu, "PPU", |ui, ppu| {
-                let PpuPermissions { display } = ppu;
-                self.show_perm_bool(ui, *display, "Display");
-            });
-            self.show_perm_bool(ui, *input, "Input");
-            self.show_perm_collapsible(ui, control, "Control", |ui, control| {
-                let ControlPermissions { dialog, pause } = control;
-                self.show_perm_bool(ui, *pause, "Pause");
-                self.show_perm_bool(ui, *dialog, "Dialog");
-            });
-        });
-
-        self.force_show_perm_collapsible(ui, external, "External", |ui, external| {
-            let ExternalPermissions { filesystem, http } = external;
-
-            self.show_perm_bool(ui, *http, "HTTP");
-            self.show_perm_collapsible(ui, filesystem, "Filesystem", |ui, fs| {
-                let FileSystemPermissions { read, write } = fs;
-                self.show_perm_bool(ui, *read, "Read");
-                self.show_perm_collapsible(ui, write, "Write", |ui, write| match write {
-                    AllOr::All => {
-                        ui.label(RichText::new("all").strong());
-                    }
-                    AllOr::Inner(files) => {
-                        for (file, options) in files.files.iter() {
-                            ui.horizontal(|ui| {
-                                ui.spacing_mut().item_spacing.x = 0.;
-                                ui.label(RichText::new(format!("{file:?}")).monospace());
-                                let label = match options {
-                                    FileWriteOptions::NewOnly => "NewOnly",
-                                    FileWriteOptions::CanOverwrite { create, mode } => &format!(
-                                        ": {}{mode:?}",
-                                        if *create { "Create + " } else { "" }
-                                    ),
-                                };
-                                ui.label(label);
-                            });
-                        }
-                    }
-                })
-            });
-        });
-
-        ui.add_space(8.0);
-        ui.horizontal(|ui| {
-            ui.checkbox(&mut self.show_none, "Show 'none' fields");
-        });
-
-        ui.separator();
-        ui.add_space(16.0);
-
-        ui.horizontal(|ui| {
-            if ui.button("Grant requested permissions").clicked() {
-                self.allow_all = true;
-                close(ui);
-            }
-            if ui.button("Cancel plugin execution").clicked() {
-                self.allow_all = false;
-                close(ui);
-            }
-        });
     }
 }
